Use get_type_hints to resolve to actual type instead of string version. (#15)
fix: use get_type_hints to resolve to actual type instead of string version

If you use from __future__ import annotations the types on the dataclass fields are by default strings and not the actual types. This behaviour will become the standard in python 3.11. Here we use the function get_type_hints to resolve to actual types.
Co-authored-by: Raphael Cristal <raphael.cristal@limehome.de>