Only ignoring direct descendants of ABC (#27)
Fixed logical issue, all descendants of class inheriting from ABC are also subclasses of ABC, meaning that if we change their schemas it is not picked up. However, only direct descendants are a part of ABC's subclasses, therefore we can check there as well. This means that classes which inherit from a class marked as abstract MUST implement any required fields.

Co-authored-by: Alex Martynek <alexander.martynek@limehome.de>